requirements.txt
kafka-python
python-dotenv
jmxquery
